+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2024-02-23
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Footer and Header common components.
+
 ## [1.4.0] - 2024-02-23
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Bootstrap 5.